backend/requirements.txt
fastapi
uvicorn[standard]
python-whois
tldextract
dnspython
requests
beautifulsoup4
python-dotenv
SQLAlchemy